pyproject.toml
[project]
name = "reprep-z7"
version = "7.2"
authors = [{name = "Andrea Censi", email = "AndreaCensi@users.noreply.github.com"}]
dependencies = [
    "Pillow",
    "PyContracts3>=7",
    "docutils",
    "ipython",
    "matplotlib",
    "numpy",
    "scipy",
    "setuptools",
    "typing_extensions",
    "zuper-commons-z7",
    ]

[tool]
[tool.poetry]
name = "reprep-z7"
version = "7.2"
description = "report reproducible research tables html latex"
authors = ["Andrea Censi <AndreaCensi@users.noreply.github.com>"]
packages = [
    {from = "src", include = "reprep"},
    {from = "src", include = "reprep/demos"},
    {from = "src", include = "reprep/graphics"},
    {from = "src", include = "reprep/output"},
    {from = "src", include = "reprep/output/hdf"},
    {from = "src", include = "reprep/output/ipn"},
    {from = "src", include = "reprep/plot_utils"},
    {from = "src", include = "reprep/report_utils"},
    {from = "src", include = "reprep/report_utils/statistics"},
    {from = "src", include = "reprep/report_utils/statistics/library"},
    {from = "src", include = "reprep/report_utils/statistics/structures"},
    {from = "src", include = "reprep/report_utils/statistics/tables"},
    {from = "src", include = "reprep/report_utils/storing"},
    {from = "src", include = "reprep/utils"},
    {from = "src", include = "reprep_tables"},
    ]

[tool.setuptools]
packages = [
    "reprep",
    "reprep.demos",
    "reprep.graphics",
    "reprep.output",
    "reprep.output.hdf",
    "reprep.output.ipn",
    "reprep.plot_utils",
    "reprep.report_utils",
    "reprep.report_utils.statistics",
    "reprep.report_utils.statistics.library",
    "reprep.report_utils.statistics.structures",
    "reprep.report_utils.statistics.tables",
    "reprep.report_utils.storing",
    "reprep.utils",
    "reprep_tables",
    ]
package-dir = {"" = "src"}
